# openapi/lucky-wheel-server.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lucky Wheel Server API",
    "version": "2.1.0",
    "description": "Production integration contract for Lucky Wheel session launch, player APIs, spin execution, and realtime updates."
  },
  "servers": [
    {
      "url": "https://api.luckywheel.example.com/api/v2",
      "description": "Production"
    },
    {
      "url": "https://sandbox-api.luckywheel.example.com/api/v2",
      "description": "Sandbox"
    }
  ],
...
  "paths": {
    "/player/session/launch": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Create player launch session",
        "description": "Merchant-signed endpoint that creates a Lucky Wheel player session for the current live admin-managed event and returns player bearer tokens.",
        "security": [
          {
            "MerchantIdHeader": [],
            "MerchantTimestampHeader": [],
            "MerchantNonceHeader": [],
            "MerchantSignatureHeader": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          },
          {
            "$ref": "#/components/parameters/Tracestate"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaunchSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Player session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaunchSessionEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/config/localization": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Get localization configuration",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          }
        ],
        "responses": {
          "200": {
            "description": "Localization resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalizationEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/events/current": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Get current live event bootstrap",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          }
        ],
        "responses": {
          "200": {
            "description": "Current event bootstrap",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventBootstrapEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "List events",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated event statuses."
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/events/{eventId}": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Get event detail",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          }
        ],
        "responses": {
          "200": {
            "description": "Event detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/events/{eventId}/leaderboard": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Get leaderboard",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/events/{eventId}/prizes": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Get prize tiers",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          }
        ],
        "responses": {
          "200": {
            "description": "Prize data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrizeEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/events/{eventId}/me": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Get current player summary",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          }
        ],
        "responses": {
          "200": {
            "description": "Player summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerSummaryEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/events/{eventId}/me/history": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Get player spin history",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spin history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpinHistoryEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/me/lucky-wheel/history/events": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Get archived event history",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archived event history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchivedHistoryEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/events/{eventId}/eligibility": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Get daily spin eligibility",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          },
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          }
        ],
        "responses": {
          "200": {
            "description": "Eligibility decision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EligibilityEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/BusinessDenied"
          },
          "503": {
            "$ref": "#/components/responses/UpstreamUnavailable"
          }
        }
      }
    },
    "/events/{eventId}/realtime": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "summary": "Subscribe to realtime event stream",
        "description": "Returns Server-Sent Events with media type text/event-stream.",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream opened",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "example": "event: leaderboard:top30\\ndata: {\\\"eventId\\\":\\\"evt_2026_march\\\"}\\n\\n"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/spins": {
      "post": {
        "tags": [
          "Player"
        ],
        "summary": "Execute a spin",
        "security": [
          {
            "PlayerBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpinRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spin executed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpinEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/BusinessDenied"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/UpstreamUnavailable"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "PlayerBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Player access token."
      },
      "MerchantIdHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Merchant-Id"
      },
      "MerchantTimestampHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Timestamp"
      },
      "MerchantNonceHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Nonce"
      },
      "MerchantSignatureHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Signature",
        "description": "HEX(HMAC_SHA256(merchantSecret, canonicalRequest))."
      }
    },
    "parameters": {
      "EventId": {
        "name": "eventId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "XRequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "Traceparent": {
        "name": "traceparent",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "Tracestate": {
        "name": "tracestate",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Conflict": {
        "description": "Replay or idempotency conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "BusinessDenied": {
        "description": "Business rule denied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "headers": {
          "X-RateLimit-Limit": {
            "schema": {
              "type": "integer"
            }
          },
          "X-RateLimit-Remaining": {
            "schema": {
              "type": "integer"
            }
          },
          "X-RateLimit-Reset": {
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "UpstreamUnavailable": {
        "description": "Upstream dependency unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
      "Meta": {
        "type": "object",
        "required": [
          "requestId",
          "traceId",
          "timestamp",
          "version"
        ],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "data": {
            "type": "object",
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ErrorObject"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "DeviceInfo": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          }
        }
      },
      "LaunchSessionRequest": {
        "type": "object",
        "required": [
          "merchantPlayerId"
        ],
        "properties": {
          "merchantPlayerId": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceInfo"
          }
        }
      },
      "LaunchSessionData": {
        "type": "object",
        "required": [
          "sessionId",
          "launchUrl",
          "accessToken",
          "refreshToken",
          "expiresAt"
        ],
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "launchUrl": {
            "type": "string",
            "format": "uri"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "LocalizationData": {
        "type": "object",
        "required": [
          "requestedLocale",
          "resolvedLocale",
          "supportedLocales"
        ],
        "properties": {
          "requestedLocale": {
            "type": "string"
          },
          "resolvedLocale": {
            "type": "string"
          },
          "supportedLocales": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocaleItem"
            }
          }
        }
      },
...
      "EventBootstrapData": {
        "type": "object",
        "required": [
          "event",
          "player"
        ],
        "properties": {
          "event": {
            "$ref": "#/components/schemas/EventSummary"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerSummary"
          }
        }
      },
...
      "SpinRequest": {
        "type": "object",
        "required": [
          "eventId"
        ],
        "properties": {
          "eventId": {
            "type": "string"
          }
        }
      },
...
      "LaunchSessionEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/LaunchSessionData"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
...
      "LocalizationEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/LocalizationData"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "EventBootstrapEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/EventBootstrapData"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "EventListEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/EventListData"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "EventEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/EventSummary"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "LeaderboardEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/LeaderboardData"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "PrizeEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/PrizeListData"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "PlayerSummaryEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/PlayerSummary"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "SpinHistoryEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/SpinHistoryData"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "ArchivedHistoryEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ArchivedEventHistoryData"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "EligibilityEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/EligibilityData"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "SpinEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/SpinResult"
          },
          "error": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      }
    }
  }
}